Fix x/X on closing delimiters
Also fix adjustment of copied text based on deleted characters.
Previously, the region beginning was never adjusted and the end was not
always adjusted correctly. Use markers, so that the adjustment happens
automatically (no need to check on the number of characters deleted
before the region beginning or the number of characters deleted in the
region).

Fixes #47.